finished initial mapping code--no localization yet
a couple of bugs still there... meh. 1. check out the comment in the
create_line_from_vector function, line 106 of geometry.js--I remember
putting it there for a reason (what reason, though?), but the gridtracer
seems to screw up on some grid edges if it's uncommented. 2. not sure if
this is a bug, but flat (flat relative to the size of grid cells)
objects are sometimes systematically forgotten.